Switch to newer timeout syntax in tests; move TestPane to its own file
